split-files split out each class into its own file in the src dir
