+++ package.json
@@ -4,7 +4,7 @@
   "description": "A uility for building conversational responses in DialogFlow fufillments",
   "main": "index.js",
   "scripts": {
-    "test": "eslint index.js"
+    "test": "eslint *.js src"
   },
   "author": "tmtek",
   "license": "ISC",
